Refactored `EchoData.update_platform` to improve readability (#1209)
* refactored update_platform method

* removed extra print statements
